Refactor: websocket 연결 부분 분리 완료
